+++ src/pyscaffold/utils.py
@@ -3,6 +3,7 @@
 Miscellaneous utilities and tools
 """
 
+import errno
 import functools
 import keyword
 import logging
@@ -236,7 +237,7 @@ def create_file(path, content, pretend=False):
             but operation is logged.
     """
     if not pretend:
-        with open(str(path), 'w') as fh:
+        with open(path, 'w', encoding='utf-8') as fh:
             fh.write(content)
 
     logger.report('create', path)
@@ -335,3 +336,128 @@ def setdefault(dict_ref, key, value):
         return dict_ref
     dict_ref[key] = value
     return dict_ref
+
+
+def localize_path(path_string):
+    """Localize path for Windows, Unix, i.e. / or \
+
+    Args:
+        path_string (str): path using /
+
+    Returns:
+        str: path depending on OS
+    """
+    return str(Path(path_string))
+
+
+#: Windows-specific error code indicating an invalid pathname.
+ERROR_INVALID_NAME = 123
+
+
+def is_pathname_valid(pathname):
+    """Check if a pathname is valid
+
+    Code by Cecil Curry from StackOverflow
+
+    Args:
+        pathname (str): string to validate
+
+    Returns:
+        `True` if the passed pathname is a valid pathname for the current OS;
+        `False` otherwise.
+    """
+    # If this pathname is either not a string or is but is empty, this pathname
+    # is invalid.
+    try:
+        if not isinstance(pathname, str) or not pathname:
+            return False
+
+        # Strip this pathname's Windows-specific drive specifier (e.g., `C:\`)
+        # if any. Since Windows prohibits path components from containing `:`
+        # characters, failing to strip this `:`-suffixed prefix would
+        # erroneously invalidate all valid absolute Windows pathnames.
+        _, pathname = os.path.splitdrive(pathname)
+
+        # Directory guaranteed to exist. If the current OS is Windows, this is
+        # the drive to which Windows was installed (e.g., the "%HOMEDRIVE%"
+        # environment variable); else, the typical root directory.
+        root_dirname = os.environ.get('HOMEDRIVE', 'C:') \
+            if sys.platform == 'win32' else os.path.sep
+        assert os.path.isdir(root_dirname)  # ...Murphy and her ironclad Law
+
+        # Append a path separator to this directory if needed.
+        root_dirname = root_dirname.rstrip(os.path.sep) + os.path.sep
+
+        # Test whether each path component split from this pathname is valid or
+        # not, ignoring non-existent and non-readable path components.
+        for pathname_part in pathname.split(os.path.sep):
+            try:
+                os.lstat(root_dirname + pathname_part)
+            # If an OS-specific exception is raised, its error code
+            # indicates whether this pathname is valid or not. Unless this
+            # is the case, this exception implies an ignorable kernel or
+            # filesystem complaint (e.g., path not found or inaccessible).
+            #
+            # Only the following exceptions indicate invalid pathnames:
+            #
+            # * Instances of the Windows-specific "WindowsError" class
+            #   defining the "winerror" attribute whose value is
+            #   "ERROR_INVALID_NAME". Under Windows, "winerror" is more
+            #   fine-grained and hence useful than the generic "errno"
+            #   attribute. When a too-long pathname is passed, for example,
+            #   "errno" is "ENOENT" (i.e., no such file or directory) rather
+            #   than "ENAMETOOLONG" (i.e., file name too long).
+            # * Instances of the cross-platform "OSError" class defining the
+            #   generic "errno" attribute whose value is either:
+            #   * Under most POSIX-compatible OSes, "ENAMETOOLONG".
+            #   * Under some edge-case OSes (e.g., SunOS, *BSD), "ERANGE".
+            except OSError as exc:
+                if hasattr(exc, 'winerror'):
+                    if exc.winerror == ERROR_INVALID_NAME:
+                        return False
+                elif exc.errno in {errno.ENAMETOOLONG, errno.ERANGE}:
+                    return False
+    # If a "TypeError" exception was raised, it almost certainly has the
+    # error message "embedded NUL character" indicating an invalid pathname.
+    except TypeError:
+        return False
+    # If no exception was raised, all path components and hence this
+    # pathname itself are valid. (Praise be to the curmudgeonly python.)
+    else:
+        return True
+    # If any other exception was raised, this is an unrelated fatal issue
+    # (e.g., a bug). Permit this exception to unwind the call stack.
+    #
+    # Did we mention this should be shipped with Python already?
+
+
+def on_ro_error(func, path, exc_info):
+    """Error handler for ``shutil.rmtree``.
+
+    If the error is due to an access error (read only file)
+    it attempts to add write permission and then retries.
+
+    If the error is for another reason it re-raises the error.
+
+    Usage : ``shutil.rmtree(path, onerror=onerror)``
+
+    Args:
+        func (callable): function which raised the exception
+        path (str): path passed to `func`
+        exc_info (tuple of str): exception info returned by sys.exc_info()
+    """
+    import stat
+    if not os.access(path, os.W_OK):
+        # Is the error an access error ?
+        os.chmod(path, stat.S_IWUSR)
+        func(path)
+    else:
+        raise
+
+
+def rm_rf(path):
+    """Remove a path by all means like `rm -rf` in Linux.
+
+    Args (str): Path to remove:
+    """
+    shutil.rmtree(path, onerror=on_ro_error)