+++ tests/test_shell.py
@@ -34,7 +34,7 @@ def test_shell_command_error2exit_decorator():
 
 
 def test_command_exists():
-    assert shell.command_exists(basename(sys.executable))
+    assert shell.command_exists('tar')
     assert not shell.command_exists('ldfgyupmqzbch174')